Rename the new reporting package.
