+++ jplag/src/main/java/de/jplag/reporting/Report.java
@@ -1,4 +1,4 @@
-package de.jplag.reporting2;
+package de.jplag.reporting;
 
 import java.util.List;
