+++ jplag/src/main/java/de/jplag/reporting/reportobject/model/FilesOfSubmission.java
@@ -1,4 +1,4 @@
-package de.jplag.reporting2.reportobject.model;
+package de.jplag.reporting.reportobject.model;
 
 import java.util.List;
